Quiet some narrowing conversion warnings from VS2015

+++ src/p_user.cpp
@@ -3420,7 +3420,7 @@ void P_PlayerThink (player_t *player, ticcmd_t *pCmd)
 		{
 			// A negative scale is used to prevent G_AddViewAngle/G_AddViewPitch
 			// from scaling with the FOV scale.
-			desired *= fabs(player->ReadyWeapon->FOVScale);
+			desired *= fabsf(player->ReadyWeapon->FOVScale);
 		}
 		if (player->FOV != desired)
 	{
